v4.1.2.0 Refactor `/unit_tests` to use Pytest. (#808)
