+++ config/params_template.env
@@ -31,20 +31,20 @@ export mask_leveed_area_toggle=True # Toggle to mask levee-protected areas from 
 
 #### estimating bankfull stage in SRCs ####
 export src_bankfull_toggle="True" # Toggle to run identify_bankfull routine (True=on; False=off)
-export bankfull_flows_file="data/inputs/rating_curve/bankfull_flows/nwm_high_water_threshold_cms.csv" # input file location with nwm feature_id and recurrence flow values
+export bankfull_flows_file="/data/inputs/rating_curve/bankfull_flows/nwm_high_water_threshold_cms.csv" # input file location with nwm feature_id and recurrence flow values
 
 #### applying channel/overbank subdivision routine to SRCs ####
 export src_subdiv_toggle="True" # Toggle to run composite roughness src routine (True=on; False=off)
 export vrough_suffix="" # text to append to output log and hydrotable file names (use for testing/debugging)
-export vmann_input_file="data/inputs/rating_curve/variable_roughness/mannings_global_06_12.csv" # input file location with nwm feature_id and channel roughness and overbank roughness attributes
+export vmann_input_file="/data/inputs/rating_curve/variable_roughness/mannings_global_06_12.csv" # input file location with nwm feature_id and channel roughness and overbank roughness attributes
 
 #### apply SRC adjustments using USGS rating curve database (calibration db) ####
 export src_adjust_usgs="True" # Toggle to run src adjustment routine (True=on; False=off)
-export nwm_recur_file="data/inputs/rating_curve/nwm_recur_flows/nwm21_17C_recurrence_flows_cfs.csv" # input file location with nwm feature_id and recurrence flow values
+export nwm_recur_file="/data/inputs/rating_curve/nwm_recur_flows/nwm21_17C_recurrence_flows_cfs.csv" # input file location with nwm feature_id and recurrence flow values
 
 #### apply SRC adjustments using observed FIM/flow point database ####
 export src_adjust_spatial="True" # Toggle to run src adjustment routine (True=on; False=off)
-export fim_obs_pnt_data="data/inputs/rating_curve/water_edge_database/usgs_nws_benchmark_points_cleaned.gpkg"
+export fim_obs_pnt_data="/data/inputs/rating_curve/water_edge_database/usgs_nws_benchmark_points_cleaned.gpkg"
 #### path to env file with sensitive paths for accessing postgres database ####
 export CALB_DB_KEYS_FILE="/data/config/calb_db_keys.env"
