+++ unit_tests/__template.json
@@ -1,11 +1,11 @@
 {
 	"valid_data":
 	{
-		"some param name (suggested to match the method param name": "some value",
+		"some param name (suggested to match the method param name)": "some value",
 		"another one and as many lines as you need": "some value",
 		"for 'None', you can just leave this field out, but sometimes it seems like jsut passing the value of None seems to work (not validated) ": "some value",
 		"for numerics, just put the number": 20,
-	   "for lists/collections: ": ["value 1", "value 2"],
+	    "for lists/collections: ": ["value 1", "value 2"],
 		"inline json comments don't work (let us know if you figure it out)": "sorry"
 	}
 }
