+++ unit_tests/aggregate_branch_lists_params.json
@@ -1,8 +1,8 @@
 {
 	"valid_data":
 	{
-		"output_dir": "/outputs/rob_wb_3/",
+		"output_dir": "/outputs/fim_unit_test_data_do_not_remove/",
 		"file_name": "branch_ids.csv",
-		"output_file_name": "/outputs/rob_wb_3/test.csv"
+		"output_file_name": "/outputs/fim_unit_test_data_do_not_remove/test.csv"
 	}
 }
