+++ unit_tests/check_unit_errors_params.json
@@ -1,11 +1,11 @@
 {
 	"valid_data":
 	{
-		"fim_dir": "/outputs/gms_example_unit_tests"
+		"fim_dir": "/data/outputs/fim_unit_test_data_do_not_remove"
 	},
 
 	"invalid_path":
 	{
-		"fim_dir": "/outputs/gms_example_unit_tests_not_not_valid"
+		"fim_dir": "/data/outputs/gms_example_unit_tests_not_not_valid"
 	}	
 }
