+++ unit_tests/clip_vectors_to_wbd_params.json
@@ -1,12 +1,14 @@
 {
 	"valid_data": {
+		"outputRunDataDir":  "/data/outputs/fim_unit_test_data_do_not_remove",
 		"subset_nwm_lakes": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/nwm_lakes_proj_subset.gpkg",
 		"subset_nwm_streams": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/nwm_subset_streams.gpkg",
 		"hucCode": "05030104",
-		"subset_nhd_headwaters": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/nhd_headwater_points_subset.gpkg",
+		"subset_nwm_headwaters": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/nhd_headwater_points_subset.gpkg",
 		"wbd_buffer_filename": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/wbd_buffered.gpkg",
 		"wbd_filename": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/wbd.gpkg",
 		"dem_filename": "/data/inputs/3dep_dems/10m_5070/fim_seamless_3dep_dem_10m_5070.vrt",
+		"dem_domain" : "/data/inputs/3dep_dems/10m_5070/HUC6_dem_domain.gpkg",
 		"nwm_lakes": "/data/inputs/nwm_hydrofabric/nwm_lakes.gpkg",
 		"nwm_catchments": "/data/inputs/nwm_hydrofabric/nwm_catchments.gpkg",
 		"subset_nwm_catchments": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/nwm_catchments_proj_subset.gpkg",
@@ -14,12 +16,11 @@
 		"landsea": "/data/inputs/landsea/water_polygons_us.gpkg",
 		"nwm_streams": "/data/inputs/nwm_hydrofabric/nwm_flows.gpkg",
 		"subset_landsea": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/LandSea_subset.gpkg",
-		"nhd_headwaters": "/data/inputs/nhdplus_vectors_aggregate/agg_nhd_headwaters_adj.gpkg",
+		"nwm_headwaters": "/data/inputs/nhdplus_vectors_aggregate/agg_nhd_headwaters_adj.gpkg",
 		"subset_nld_lines": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/nld_subset_levees.gpkg",
-		"great_lakes": "/data/inputs/landsea/gl_water_polygons.gpkg",
-		"lake_buffer_distance": 20,
 		"wbd_buffer_distance": 5000,
 		"levee_protected_areas": "/data/inputs/nld_vectors/Levee_protected_areas.gpkg",
-		"subset_levee_protected_areas": "/data/outputs/rob_test_2/05030104/LeveeProtectedAreas_subset.gpkg"
+		"subset_levee_protected_areas": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/LeveeProtectedAreas_subset.gpkg"
 	}
 }
+
