--- unit_tests/generate_branch_list_unittests.py
@@ -1,59 +1,0 @@
-#!/usr/bin/env python3
-
-import inspect
-import os
-import sys
-
-import json
-import warnings
-import unittest
-
-sys.path.append('/foss_fim/unit_tests/')
-from unit_tests_utils import FIM_unit_test_helpers as ut_helpers
-
-# importing python folders in other directories
-sys.path.append('/foss_fim/src/')   # *** update your folder path here if required ***
-import generate_branch_list as src
-
-class test_Generate_branch_list(unittest.TestCase):
-
-    '''
-    Allows the params to be loaded one and used for all test methods
-    '''
-    @classmethod
-    def setUpClass(self):
-
-        warnings.simplefilter('ignore')
-        params_file_path = ut_helpers.get_params_filename(__file__)
-        with open(params_file_path) as params_file:
-            self.params = json.load(params_file)
-
-
-    def test_Generate_branch_list_success(self):
-
-        #global params_file
-        
-        params = self.params["valid_data"].copy()  #update "valid_data" value if you need to (aka.. more than one node)
-
-        src.generate_branch_list(stream_network_dissolved = params["stream_network_dissolved"],
-                                 branch_id_attribute = params["branch_id_attribute"],
-                                 output_branch_list_file = params["output_branch_list_file"])
-       
-       
-        print(f"Test Success: {inspect.currentframe().f_code.co_name}")
-        print("*************************************************************")
-
-
-if __name__ == '__main__':
-
-    script_file_name = os.path.basename(__file__)
-
-    print("*****************************")
-    print(f"Start of {script_file_name} tests")
-    print()
-   
-    unittest.main()
-    
-    print()    
-    print(f"End of {script_file_name} tests")
-
