+++ unit_tests/gms/derive_level_paths_params.json
@@ -1,6 +1,7 @@
 {
 	"valid_data":
 	{
+		"outputRunDataDir" : "/data/outputs/fim_unit_test_data_do_not_remove/",
 		"in_stream_network": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/nwm_subset_streams.gpkg",
 		"out_stream_network": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/nwm_subset_streams_levelPaths.gpkg",
 		"branch_id_attribute": "levpa_id",
@@ -11,21 +12,7 @@
 		"branch_inlets_outfile": "/data/outputs/fim_unit_test_data_do_not_remove/05030104/nwm_subset_streams_levelPaths_dissolved_headwaters.gpkg",
 		"reach_id_attribute": "ID",
 		"verbose": true
-	},
-	
-	"dropped_stream_orders_no_branches_remaining":
-	{
-		"in_stream_network": "/data/outputs/fim_unit_test_data_do_not_remove/02030201/nwm_subset_streams.gpkg",
-		"out_stream_network": "/data/outputs/fim_unit_test_data_do_not_remove/02030201/nwm_subset_streams_levelPaths.gpkg",
-		"branch_id_attribute": "levpa_id",
-		"out_stream_network_dissolved": "/data/outputs/fim_unit_test_data_do_not_remove/02030201/nwm_subset_streams_levelPaths.gpkg",
-		"headwaters_outfile": "/data/outputs/fim_unit_test_data_do_not_remove/02030201/nwm_headwaters.gpkg",
-		"catchments": "/data/outputs/fim_unit_test_data_do_not_remove/02030201/nwm_catchments_proj_subset.gpkg",
-		"catchments_outfile": "/data/outputs/fim_unit_test_data_do_not_remove/02030201/nwm_catchments_proj_subset_levelPaths.gpkg",
-		"branch_inlets_outfile": "/data/outputs/fim_unit_test_data_do_not_remove/02030201/nwm_subset_streams_levelPaths_dissolved_headwaters.gpkg",
-		"reach_id_attribute": "ID",
-		"verbose": true
-	}	
+	}
 }
