+++ unit_tests/shared_functions_params.json
@@ -1,15 +1,17 @@
 {
 	"append_append_id_to_file_name_single_identifier_valid":
 	{
-		"file_name": "/output/fim_unit_test_data_do_not_remove/02020005/branches/3246000003/rem_zeroed_masked_3246000003.tif",
+		"outputRunDataDir": "/output/fim_unit_test_data_do_not_remove/",
+		"file_name": "02020005/branches/3246000003/rem_zeroed_masked_3246000003.tif",
 		"identifier": "8888",
-		"expected_output": "/output/fim_unit_test_data_do_not_remove/02020005/branches/3246000003/rem_zeroed_masked_3246000003_8888.tif"
+		"expected_output": "02020005/branches/3246000003/rem_zeroed_masked_3246000003_8888.tif"
 	},
 
 	"append_append_id_to_file_name_identifier_list_valid":
 	{
-		"file_name": "/output/fim_unit_test_data_do_not_remove/02020005/branches/3246000003/rem_zeroed_masked_3246000003.tif",
+		"outputRunDataDir": "/output/fim_unit_test_data_do_not_remove/",
+		"file_name": "02020005/branches/3246000003/rem_zeroed_masked_3246000003.tif",
 		"identifier": ["7777", "8888"],
-		"expected_output": "/output/fim_unit_test_data_do_not_remove/02020005/branches/3246000003/rem_zeroed_masked_3246000003_7777_8888.tif"
+		"expected_output": "02020005/branches/3246000003/rem_zeroed_masked_3246000003_7777_8888.tif"
 	}	
 }
