+++ unit_tests/split_flows_params.json
@@ -1,6 +1,7 @@
 {
 	"valid_data":
 	{
+		"outputRunDataDir" : "/data/outputs/fim_unit_test_data_do_not_remove/",
 		"max_length": 1500,
 		"slope_min": 0.001,
 		"lakes_buffer_input": 20,
@@ -10,6 +11,6 @@
 		"split_points_filename": "/data/outputs/fim_unit_test_data_do_not_remove/02020005/branches/3246000005/demDerived_reaches_split_points_3246000005.gpkg",
 		"wbd8_clp_filename": "/data/outputs/fim_unit_test_data_do_not_remove/02020005/wbd8_clp.gpkg",
 		"lakes_filename": "/data/outputs/fim_unit_test_data_do_not_remove/02020005/nwm_lakes_proj_subset.gpkg",
-		"drop_stream_orders": true
+		"nwm_streams_filename": "/data/outputs/fim_unit_test_data_do_not_remove/02020005/nwm_subset_streams_levelPaths.gpkg"
 	}
 }
