+++ unit_tests/unit_tests_utils.py
@@ -10,7 +10,7 @@ class FIM_unit_test_helpers(object):
     def get_params_filename(unit_test_file_name):
        
         unittest_file_name = os.path.basename(unit_test_file_name)
-        params_file_name = unittest_file_name.replace("_unittests.py", "_params.json")
+        params_file_name = unittest_file_name.replace("_test.py", "_params.json")
         params_file_path = os.path.join(os.path.dirname(unit_test_file_name), params_file_name)
         
         if (not os.path.exists(params_file_path)):
@@ -18,3 +18,9 @@ class FIM_unit_test_helpers(object):
         
         return params_file_path
     
+    @staticmethod
+    def json_concat(obj, keyone, keytwo):
+        
+        concat_string = obj[keyone] + obj[keytwo] 
+        
+        return concat_string
